+++ lib/devices/device_descriptors/concept2_row_erg.dart
@@ -1,0 +1,14 @@
+import '../device_fourcc.dart';
+import 'concept2_erg.dart';
+
+class Concept2RowErg extends Concept2Erg {
+  Concept2RowErg()
+      : super(
+          deviceSportDescriptors[concept2RowerFourCC]!.defaultSport,
+          deviceSportDescriptors[concept2RowerFourCC]!.isMultiSport,
+          concept2RowerFourCC,
+        );
+
+  @override
+  Concept2RowErg clone() => Concept2RowErg();
+}