+++ lib/devices/device_descriptors/kayak_first_descriptor.dart
@@ -1,6 +1,7 @@
 import 'dart:collection';
 import 'dart:convert';
 import 'package:flutter/foundation.dart';
+import 'package:flutter/material.dart';
 import 'package:flutter_blue_plus/flutter_blue_plus.dart';
 import 'package:get/get.dart';
 import 'package:pref/pref.dart';
@@ -11,6 +12,8 @@ import '../../preferences/athlete_body_weight.dart';
 import '../../preferences/block_signal_start_stop.dart';
 import '../../preferences/kayak_first_display_configuration.dart';
 import '../../preferences/log_level.dart';
+import '../../ui/models/progress_state.dart';
+import '../../ui/parts/pre_measurement_progress.dart';
 import '../../utils/constants.dart';
 import '../../utils/logging.dart';
 import '../gatt/ftms.dart';
@@ -47,15 +50,17 @@ class KayakFirstDescriptor extends DeviceDescriptor {
   static const responseWatchTimeoutGuard = Duration(milliseconds: responseWatchTimeoutMs + 250);
   static const commandShortDelayMs = 500; // ms
   static const commandShortDelay = Duration(milliseconds: commandShortDelayMs);
-  static const commandLongDelayMs = 2000; // ms
+  static const commandLongDelayMs = 1000; // ms
   static const commandLongDelay = Duration(milliseconds: commandLongDelayMs);
   static const commandExtraLongDelayMs = 5000; // ms
   static const commandExtraLongDelay = Duration(milliseconds: commandExtraLongDelayMs);
   static const commandExtraLongTimeoutGuard = Duration(milliseconds: commandExtraLongDelayMs + 250);
+  static const progressBarCompletionTime = 6000; // ms
   static const maxWaitIterations = 3;
   static const List<int> validFlags = [
     resetByte + 0x0D * 256,
     resetByte + separator * 256,
+    resetByte + 1 * 256,
     handshakeByte + separator * 256,
     configurationByte + separator * 256,
     displayConfigurationByte + separator * 256,
@@ -137,13 +142,17 @@ class KayakFirstDescriptor extends DeviceDescriptor {
 
   @override
   bool isWholePacket(List<int> data) {
-    return data.length >= 2 && data.last == 0x0A && data[data.length - 2] == 0x0D;
+    return data.length >= 2 && data.last == 0x0A && data[data.length - 2] == 0x0D ||
+        data.length == 2 && data.first == resetByte && data.last == 1;
   }
 
   @override
   bool skipPacket(List<int> data) {
-    return data.length == 2 && data.first == resetByte && data.last == 1 ||
-        data.length == 3 && data.first == dataStreamFlag && data.last == 0x0A && data[1] == 0x0D;
+    return data.length == 3 &&
+            data.first == dataStreamFlag &&
+            data.last == 0x0A &&
+            data[1] == 0x0D ||
+        data.isNotEmpty && data.map((e) => e == 0 || e == 1).every((p) => p);
   }
 
   @override
@@ -258,90 +267,101 @@ class KayakFirstDescriptor extends DeviceDescriptor {
       return;
     }
 
-    await Future.delayed(commandShortDelay);
+    progressBottomSheet("Initializing 1...", progressBarCompletionTime);
+    await Future.delayed(responseWatchDelay);
     final prefService = Get.find<BasePrefService>();
     final blockSignalStartStop =
         testing || (prefService.get<bool>(blockSignalStartStopTag) ?? blockSignalStartStopDefault);
-    // 1. Reset
+    // 1.1 Reset
     bool seenIt = false;
     int iterationCount = 0;
     currentResponses.clear();
     while (!seenIt && iterationCount < maxWaitIterations) {
       await executeControlOperation(controlPoint, blockSignalStartStop, logLevel, resetControl);
-      seenIt = await _waitForResponse(resetByte, responseWatchTimeoutMs, logLevel)
+      seenIt = await _waitForResponse(resetByte, commandExtraLongDelayMs, logLevel, true)
           .timeout(responseWatchTimeoutGuard, onTimeout: () => false);
       await Future.delayed(commandLongDelay);
       iterationCount++;
     }
 
-    await Future.delayed(commandShortDelay);
+    await Future.delayed(commandLongDelay);
+    ProgressState.optionallyCloseProgress();
+    await Future.delayed(responseWatchDelay);
+    // 1.2 Reset
+    seenIt = false;
+    iterationCount = 0;
+    currentResponses.clear();
+    progressBottomSheet("Initializing 2...", progressBarCompletionTime);
+    while (!seenIt && iterationCount < maxWaitIterations) {
+      await executeControlOperation(controlPoint, blockSignalStartStop, logLevel, resetControl);
+      seenIt = await _waitForResponse(resetByte, commandExtraLongDelayMs, logLevel, true)
+          .timeout(responseWatchTimeoutGuard, onTimeout: () => false);
+      await Future.delayed(commandLongDelay);
+      iterationCount++;
+    }
+
+    await Future.delayed(commandExtraLongDelay);
+    ProgressState.optionallyCloseProgress();
+    await Future.delayed(responseWatchDelay);
     // 2. Handshake
     seenIt = false;
     iterationCount = 0;
     currentResponses.clear();
+    progressBottomSheet("Handshake...", progressBarCompletionTime);
     while (!seenIt && iterationCount < maxWaitIterations) {
       await handshake(controlPoint, false, logLevel);
-      seenIt = await _waitForResponse(handshakeByte, responseWatchTimeoutMs, logLevel)
+      seenIt = await _waitForResponse(handshakeByte, commandExtraLongDelayMs, logLevel, true)
           .timeout(responseWatchTimeoutGuard, onTimeout: () => false);
       await Future.delayed(commandLongDelay);
       iterationCount++;
     }
 
-    await Future.delayed(commandShortDelay);
+    await Future.delayed(commandExtraLongDelay);
+    ProgressState.optionallyCloseProgress();
+    await Future.delayed(responseWatchDelay);
     // 3. Display Configuration
     seenIt = false;
     iterationCount = 0;
     currentResponses.clear();
+    progressBottomSheet("Configuration...", progressBarCompletionTime);
     while (!seenIt && iterationCount < maxWaitIterations) {
       await configureDisplay(controlPoint, logLevel);
-      seenIt = await _waitForResponse(displayConfigurationByte, commandExtraLongDelayMs, logLevel)
-          .timeout(commandExtraLongTimeoutGuard, onTimeout: () => false);
+      seenIt =
+          await _waitForResponse(displayConfigurationByte, commandExtraLongDelayMs, logLevel, true)
+              .timeout(commandExtraLongTimeoutGuard, onTimeout: () => false);
       await Future.delayed(commandLongDelay);
       iterationCount++;
     }
 
-    /*
-    await Future.delayed(commandShortDelay);
-    // X1. Prelim measurement start command
-    seenIt = false;
-    iterationCount = 0;
-    currentResponses.clear();
-    await executeControlOperation(
-        controlPoint, blockSignalStartStop, logLevel, startOrResumeControl);
-    while (!seenIt && iterationCount < maxWaitIterations) {
-      seenIt = await _waitForResponse(startStopByte, commandExtraLongDelayMs, logLevel)
-          .timeout(commandExtraLongTimeoutGuard, onTimeout: () => false);
-      await Future.delayed(commandShortDelay);
-      iterationCount++;
-    }
-
-    await Future.delayed(commandShortDelay);
-    // X2. Prelim measurement stop command
-    seenIt = false;
-    iterationCount = 0;
-    currentResponses.clear();
-    await executeControlOperation(controlPoint, blockSignalStartStop, logLevel, stopOrPauseControl);
-    while (!seenIt && iterationCount < maxWaitIterations) {
-      seenIt = await _waitForResponse(startStopByte, commandExtraLongDelayMs, logLevel)
-          .timeout(commandExtraLongTimeoutGuard, onTimeout: () => false);
-      await Future.delayed(commandShortDelay);
-      iterationCount++;
-    }
-    */
-
+    await Future.delayed(commandExtraLongDelay);
+    ProgressState.optionallyCloseProgress();
+    await Future.delayed(responseWatchDelay);
     initializedConsole = true;
   }
 
-  Future<bool> _waitForResponse(int responseByte, int timeout, int logLevel) async {
+  Future<bool> _waitForResponse(int responseByte, int timeout, int logLevel, bool withEcho) async {
     final iterationCount = timeout ~/ responseWatchDelayMs;
     int i = 0;
+    bool seenEcho = !withEcho;
+    if (withEcho) {
+      for (; i < iterationCount && !currentResponses.contains(responseByte); i++) {
+        await Future.delayed(responseWatchDelay);
+      }
+
+      final seenEcho = currentResponses.contains(responseByte);
+      Logging().log(logLevel, logLevelInfo, tag, "_waitForResponse echo", "$seenEcho");
+
+      currentResponses.clear();
+    }
+
+    i = 0;
     for (; i < iterationCount && !currentResponses.contains(responseByte); i++) {
       await Future.delayed(responseWatchDelay);
     }
 
     final seenIt = currentResponses.contains(responseByte);
     Logging().log(logLevel, logLevelInfo, tag, "_waitForResponse", "$seenIt");
-    return seenIt;
+    return seenIt || seenEcho;
   }
 
   @override