+++ lib/devices/device_factory.dart
@@ -1,7 +1,10 @@
 import '../export/fit/fit_manufacturer.dart';
 import '../utils/constants.dart';
 import 'device_descriptors/cross_trainer_device_descriptor.dart';
-import 'device_descriptors/concept2_rower.dart';
+import 'device_descriptors/concept2_bike_erg.dart';
+import 'device_descriptors/concept2_erg.dart';
+import 'device_descriptors/concept2_row_erg.dart';
+import 'device_descriptors/concept2_ski_erg.dart';
 import 'device_descriptors/cycling_power_meter_descriptor.dart';
 import 'device_descriptors/cycling_speed_and_cadence_descriptor.dart';
 import 'device_descriptors/device_descriptor.dart';
@@ -310,7 +313,17 @@ class DeviceFactory {
       case cscSensorBasedPaddleFourCC:
         return DeviceFactory.getCSCBasedPaddler();
       case concept2RowerFourCC:
-        return Concept2Rower();
+        return Concept2RowErg();
+      case concept2SkiFourCC:
+        return Concept2SkiErg();
+      case concept2BikeFourCC:
+        return Concept2BikeErg();
+      case concept2ErgFourCC:
+        return Concept2Erg(
+          deviceSportDescriptors[concept2ErgFourCC]!.defaultSport,
+          deviceSportDescriptors[concept2ErgFourCC]!.isMultiSport,
+          concept2ErgFourCC,
+        );
       case merachMr667FourCC:
         return getMerachMr667();
       case virtufitUltimatePro2FourCC:
@@ -348,6 +361,9 @@ class DeviceFactory {
       case ActivityType.elliptical:
         fourCC = genericFTMSCrossTrainerFourCC;
         break;
+      case ActivityType.nordicSki:
+        fourCC = concept2SkiFourCC;
+        break;
     }
 
     return DeviceFactory.getDescriptorForFourCC(fourCC);
@@ -355,5 +371,16 @@ class DeviceFactory {
 
   static List<DeviceDescriptor> allDescriptors() {
     return [for (var fourCC in allFourCC) DeviceFactory.getDescriptorForFourCC(fourCC)];
+  }
+
+  static List<String> getSportChoices(String fourCC) {
+    if (fourCC == kayakFirstFourCC) {
+      return paddleSports;
+    } else if (fourCC == concept2ErgFourCC) {
+      return c2Sports;
+    }
+
+    // KayakPro
+    return waterSports;
   }
 }