+++ lib/devices/gadgets/c2_additional_status2.dart
@@ -6,8 +6,8 @@ import '../metric_descriptors/short_metric_descriptor.dart';
 import 'complex_sensor.dart';
 
 class C2AdditionalStatus2 extends ComplexSensor {
-  static const serviceUuid = c2RowingPrimaryServiceUuid;
-  static const characteristicUuid = c2RowingAdditionalStatus2Uuid;
+  static const serviceUuid = c2ErgPrimaryServiceUuid;
+  static const characteristicUuid = c2ErgAdditionalStatus2Uuid;
 
   static const expectedDataPacketLength = 20;
   static const caloriesLsbByteIndex = 6;