+++ lib/preferences/sport_spec.dart
@@ -13,7 +13,8 @@ class SportSpec {
   static String sport2Sport(String sport) {
     return sport == ActivityType.kayaking ||
             sport == ActivityType.canoeing ||
-            sport == ActivityType.rowing
+            sport == ActivityType.rowing ||
+            sport == ActivityType.nordicSki
         ? SportSpec.paddleSport
         : sport;
   }