+++ lib/ui/recording.dart
@@ -74,12 +74,14 @@ import '../utils/target_heart_rate.dart';
 import '../utils/theme_manager.dart';
 import '../utils/time_zone.dart';
 import 'models/display_record.dart';
+import 'models/progress_state.dart';
 import 'models/row_configuration.dart';
 import 'parts/battery_status.dart';
 import 'parts/heart_rate_monitor_pairing.dart';
 import 'parts/kayak_first.dart';
 import 'parts/legend_dialog.dart';
 import 'parts/pick_directory.dart';
+import 'parts/pre_measurement_progress.dart';
 import 'parts/spin_down.dart';
 import 'parts/three_choices.dart';
 import 'parts/upload_portal_picker.dart';
@@ -277,12 +279,17 @@ class RecordingState extends State<RecordingScreen> {
       await _fitnessEquipment?.additionalSensorsOnDemand();
       await _fitnessEquipment?.attach();
       _startPumpingData();
-      await _fitnessEquipment?.postPumpStart();
+      if (!_uxDebug) {
+        await _fitnessEquipment?.postPumpStart();
+      }
+
       final prefService = Get.find<BasePrefService>();
       if (prefService.get<bool>(instantMeasurementStartTag) ?? instantMeasurementStartDefault) {
         WidgetsBinding.instance.addPostFrameCallback((_) async {
+          progressBottomSheet("Starting...", KayakFirstDescriptor.progressBarCompletionTime);
           await Future.delayed(KayakFirstDescriptor.commandLongDelay);
           await _startMeasurement(false);
+          ProgressState.optionallyCloseProgress();
         });
       }
 
@@ -361,15 +368,15 @@ class RecordingState extends State<RecordingScreen> {
               _graphMaxData.removeFirst();
             }
           }
-          graphData = _graphData.toList();
+          graphData = _graphData.toList(growable: false);
           if (_onStageStatisticsType == onStageStatisticsTypeAverage ||
               _onStageStatisticsType == onStageStatisticsTypeAlternating) {
-            graphAvgData = _graphAvgData.toList();
+            graphAvgData = _graphAvgData.toList(growable: false);
           }
 
           if (_onStageStatisticsType == onStageStatisticsTypeMaximum ||
               _onStageStatisticsType == onStageStatisticsTypeAlternating) {
-            graphMaxData = _graphMaxData.toList();
+            graphMaxData = _graphMaxData.toList(growable: false);
           }
         }
 
@@ -741,25 +748,33 @@ class RecordingState extends State<RecordingScreen> {
         prefService.get<String>(instantExportLocationTag) ?? instantExportLocationDefault;
     _calculateGps = prefService.get<bool>(calculateGpsTag) ?? calculateGpsDefault;
     _pointCount = min(60, size.width ~/ 2);
+    _onStageStatisticsType =
+        prefService.get<String>(onStageStatisticsTypeTag) ?? onStageStatisticsTypeDefault;
     final now = DateTime.now();
-    _graphData = _simplerUi
-        ? ListQueue<DisplayRecord>(0)
-        : ListQueue.from(List<DisplayRecord>.generate(
+    if (!_simplerUi) {
+      _graphData = ListQueue.from(List<DisplayRecord>.generate(
+          _pointCount,
+          (i) =>
+              DisplayRecord.blank(widget.sport, now.subtract(Duration(seconds: _pointCount - i)))));
+      graphData = _graphData.toList(growable: false);
+      if (_onStageStatisticsType == onStageStatisticsTypeAverage ||
+          _onStageStatisticsType == onStageStatisticsTypeAlternating) {
+        _graphAvgData = ListQueue.from(List<DisplayRecord>.generate(
             _pointCount,
             (i) => DisplayRecord.blank(
                 widget.sport, now.subtract(Duration(seconds: _pointCount - i)))));
-    _graphAvgData = _simplerUi
-        ? ListQueue<DisplayRecord>(0)
-        : ListQueue.from(List<DisplayRecord>.generate(
+        graphAvgData = _graphAvgData.toList(growable: false);
+      }
+
+      if (_onStageStatisticsType == onStageStatisticsTypeMaximum ||
+          _onStageStatisticsType == onStageStatisticsTypeAlternating) {
+        _graphMaxData = ListQueue.from(List<DisplayRecord>.generate(
             _pointCount,
             (i) => DisplayRecord.blank(
                 widget.sport, now.subtract(Duration(seconds: _pointCount - i)))));
-    _graphMaxData = _simplerUi
-        ? ListQueue<DisplayRecord>(0)
-        : ListQueue.from(List<DisplayRecord>.generate(
-            _pointCount,
-            (i) => DisplayRecord.blank(
-                widget.sport, now.subtract(Duration(seconds: _pointCount - i)))));
+        graphMaxData = _graphMaxData.toList(growable: false);
+      }
+    }
 
     if (widget.sport != ActivityType.ride) {
       final slowPace = SpeedSpec.slowSpeeds[SportSpec.sport2Sport(widget.sport)]!;
@@ -1688,7 +1703,7 @@ class RecordingState extends State<RecordingScreen> {
 
   Future<void> startStopAction() async {
     if (_measuring) {
-      if (_circuitWorkout) {
+      if (_circuitWorkout && !_uxDebug) {
         final selection = await Get.bottomSheet(
           const SafeArea(
             child: Column(