+++ lib/utils/scan_result_ex.dart
@@ -44,7 +44,7 @@ extension ScanResultEx on ScanResult {
           serviceUuids.contains(schwinnX70ServiceUuid) ||
           serviceUuids.contains(cyclingPowerServiceUuid) ||
           serviceUuids.contains(cyclingCadenceServiceUuid) ||
-          serviceUuids.contains(c2RowingPrimaryServiceUuid) ||
+          serviceUuids.contains(c2ErgPrimaryServiceUuid) ||
           serviceUuids.contains(kayakFirstServiceUuid) ||
           serviceUuids.contains(heartRateServiceUuid)) {
         return true;
@@ -159,7 +159,7 @@ extension ScanResultEx on ScanResult {
       }
     }
 
-    if (serviceUuids.contains(c2RowingPrimaryServiceUuid)) {
+    if (serviceUuids.contains(c2ErgPrimaryServiceUuid)) {
       return MachineType.rower;
     }
 
@@ -179,12 +179,14 @@ extension ScanResultEx on ScanResult {
     final loweredManufacturers =
         manufacturerNames().map((m) => m.toLowerCase()).toList(growable: false);
     for (MapEntry<String, DeviceIdentifierHelperEntry> mapEntry in deviceNamePrefixes.entries) {
-      if (multiSportFourCCs.contains(mapEntry.key)) {
+      if (multiSportFourCCs.contains(mapEntry.key) && mapEntry.key != concept2ErgFourCC) {
         continue;
       }
 
+      final lowerPostfix = mapEntry.value.deviceNameLoweredPostfix;
       for (final loweredPrefix in mapEntry.value.deviceNameLoweredPrefixes) {
         if (loweredPlatformName.startsWith(loweredPrefix) &&
+            (lowerPostfix.isEmpty || loweredPlatformName.endsWith(lowerPostfix)) &&
             (mapEntry.value.manufacturerNamePrefix.isEmpty ||
                 loweredManufacturers
                     .map((m) => m.contains(mapEntry.value.manufacturerNameLoweredPrefix))
@@ -199,6 +201,7 @@ extension ScanResultEx on ScanResult {
 
   Tuple2<Widget, Widget> getLogoAndBanner(List<MachineType> ftmsServiceDataMachineTypes,
       String deviceSport, double logoSize, double mediaWidth, ThemeManager themeManager) {
+    final loweredPlatformName = device.platformName.toLowerCase();
     if (advertisementData.serviceUuids.isNotEmpty) {
       final serviceUuids = advertisementData.uuids;
       if (serviceUuids.contains(schwinnX70ServiceUuid)) {
@@ -210,7 +213,7 @@ extension ScanResultEx on ScanResult {
         );
       }
 
-      if (serviceUuids.contains(c2RowingPrimaryServiceUuid)) {
+      if (serviceUuids.contains(c2ErgPrimaryServiceUuid)) {
         return Tuple2(
           Image.asset("assets/equipment/Concept2_logo.png",
               width: logoSize, semanticLabel: "Concept2 Logo"),
@@ -233,14 +236,32 @@ extension ScanResultEx on ScanResult {
           ),
         );
       }
+
+      if (loweredPlatformName.startsWith("stages") &&
+          (serviceUuids.contains(fitnessMachineUuid) ||
+              serviceUuids.contains(cyclingPowerServiceUuid))) {
+        return Tuple2(
+          SvgPicture.asset(
+            "assets/equipment/Stages_logo.svg",
+            width: logoSize,
+            semanticsLabel: "Stages Logo",
+          ),
+          SvgPicture.asset(
+            "assets/equipment/Stages_banner.svg",
+            height: logoSize,
+            semanticsLabel: "Stages Banner",
+          ),
+        );
+      }
     }
 
-    final loweredPlatformName = device.platformName.toLowerCase();
     final loweredManufacturers =
         manufacturerNames().map((m) => m.toLowerCase()).toList(growable: false);
     for (MapEntry<String, DeviceIdentifierHelperEntry> mapEntry in deviceNamePrefixes.entries) {
+      final lowerPostfix = mapEntry.value.deviceNameLoweredPostfix;
       for (final loweredPrefix in mapEntry.value.deviceNameLoweredPrefixes) {
         if (loweredPlatformName.startsWith(loweredPrefix) &&
+            (lowerPostfix.isEmpty || loweredPlatformName.endsWith(lowerPostfix)) &&
             (mapEntry.value.manufacturerNamePrefix.isEmpty ||
                 loweredManufacturers
                     .map((m) => m.contains(mapEntry.value.manufacturerNameLoweredPrefix))
@@ -276,7 +297,7 @@ extension ScanResultEx on ScanResult {
             return Tuple2(
               SvgPicture.asset(
                 "assets/equipment/Stages_logo.svg",
-                height: logoSize,
+                width: logoSize,
                 semanticsLabel: "Stages Logo",
               ),
               SvgPicture.asset(