+++ CHANGELOG.md
@@ -1,6 +1,14 @@
+## 1.1.152
+* Concept2 RowErg, SkiErg, BikeErg support progress
+* Kayak First ergometer support (both kayak and canoe) with UI feedback
+* Upgrade Flutter Blue Plus plugin several times
+* Fixing graph problem during measurement (syncfusion_flutter_charts v24+ issue)
+* Fixing Stages branding logo size and also the logo's and banner's color in dark mode
+* Fixing Bowflex logo SVG and also the logo's and banner's color in dark mode
+
 ## 1.1.151
 * Concept2 RowErg FTMS support
-* Kayak First ergometer support (both kayak and canoe)
+* Kayak First ergometer support progress (both kayak and canoe)
 * Upgrade Flutter Blue Plus plugin many times
 * Upgrade Flutter to channel beta 3.18.0-0.2.pre, Dart 3.3.0-174.3.beta
 * Upgrade numerous plugins many times